+++ autd3-driver/src/datagram/gain/transform2.rs
@@ -12,48 +12,53 @@ use std::collections::HashMap;
 /// Gain to transform gain data
 #[derive(Gain)]
 #[no_gain_transform]
-pub struct Transform<G: Gain + 'static, F: Fn(&Device, &Transducer, Drive) -> Drive + 'static> {
+pub struct Transform2<
+    G: Gain + 'static,
+    FT: Fn(&Transducer, Drive) -> Drive + 'static,
+    F: Fn(&Device) -> FT + 'static,
+> {
     gain: G,
     f: F,
 }
 
-pub trait IntoTransform<G: Gain> {
+pub trait IntoTransform2<G: Gain> {
     /// transform gain data
     ///
     /// # Arguments
     ///
     /// * `f` - transform function. The first argument is the device, the second is transducer, and the third is the original drive data.
     ///
-    #[deprecated(note = "Use with_transform2 instead")]
-    fn with_transform<F: Fn(&Device, &Transducer, Drive) -> Drive>(self, f: F) -> Transform<G, F>;
+    fn with_transform2<FT: Fn(&Transducer, Drive) -> Drive, F: Fn(&Device) -> FT>(
+        self,
+        f: F,
+    ) -> Transform2<G, FT, F>;
 }
 
-impl<G: Gain + 'static, F: Fn(&Device, &Transducer, Drive) -> Drive> Transform<G, F> {
+impl<G: Gain, FT: Fn(&Transducer, Drive) -> Drive, F: Fn(&Device) -> FT> Transform2<G, FT, F> {
     #[doc(hidden)]
     pub fn new(gain: G, f: F) -> Self {
         Self { gain, f }
     }
 }
 
-impl<G: Gain + 'static, F: Fn(&Device, &Transducer, Drive) -> Drive + 'static> Gain
-    for Transform<G, F>
+impl<G: Gain, FT: Fn(&Transducer, Drive) -> Drive, F: Fn(&Device) -> FT> Gain
+    for Transform2<G, FT, F>
 {
     fn calc(
         &self,
         geometry: &Geometry,
         filter: GainFilter,
     ) -> Result<HashMap<usize, Vec<Drive>>, AUTDInternalError> {
-        Ok(self
-            .gain
-            .calc(geometry, filter)?
-            .into_iter()
-            .map(|(k, v)| {
+        let src = self.gain.calc(geometry, filter)?;
+        Ok(geometry
+            .devices()
+            .map(|dev| {
+                let f = (self.f)(dev);
                 (
-                    k,
-                    v.into_iter()
-                        .enumerate()
-                        .map(|(i, d)| (self.f)(&geometry[k], &geometry[k][i], d))
-                        .collect::<Vec<_>>(),
+                    dev.idx(),
+                    dev.iter()
+                        .map(|tr| f(tr, src[&dev.idx()][tr.idx()]))
+                        .collect(),
                 )
             })
             .collect())
@@ -69,14 +74,13 @@ mod tests {
     use crate::{defined::FREQ_40K, geometry::tests::create_geometry};
 
     #[test]
-    #[allow(deprecated)]
     fn test() -> anyhow::Result<()> {
         let geometry = create_geometry(1, 249, FREQ_40K);
 
         let mut rng = rand::thread_rng();
         let d: Drive = Drive::new(Phase::new(rng.gen()), EmitIntensity::new(rng.gen()));
 
-        let gain = TestGain::null().with_transform(move |_, _, _| d);
+        let gain = TestGain::null().with_transform2(move |_| move |_, _| d);
 
         assert_eq!(
             geometry