+++ autd3/src/prelude.rs
@@ -9,15 +9,17 @@ pub use crate::{
 #[allow(deprecated)]
 pub use crate::gain::Bessel;
 
+// TODO: remove `GainTransform`, `GainCache`, `ModulationCache`, `ModulationTransform` and `RadiationPressure`
 pub use autd3_driver::{
     autd3_device::AUTD3,
     datagram::{
         Clear, ControlPoint, DebugSettings, EmulateGPIOIn, FocusSTM, ForceFan, GainCache,
         GainFilter, GainSTM, GainTransform, Group, IntoDatagramWithSegment,
         IntoDatagramWithSegmentTransition, IntoDatagramWithTimeout, IntoGainCache,
-        IntoGainTransform, IntoModulationCache, IntoModulationTransform, IntoRadiationPressure,
-        Modulation, ModulationCache, ModulationProperty, ModulationTransform, PhaseFilter,
-        PulseWidthEncoder, RadiationPressure, ReadsFPGAState, Silencer, SwapSegment, Synchronize,
+        IntoGainTransform, IntoGainTransform2, IntoModulationCache, IntoModulationTransform,
+        IntoRadiationPressure, Modulation, ModulationCache, ModulationProperty,
+        ModulationTransform, PhaseFilter, PulseWidthEncoder, RadiationPressure, ReadsFPGAState,
+        Silencer, SwapSegment, Synchronize,
     },
     defined::{deg, kHz, mm, rad, Hz, PI},
     error::AUTDInternalError,